fix: update parsing from parseInt to parseFloat for numeric values across multiple components

+++ src/lib/qti-components/qti-response-processing/qti-expression/qti-product/qti-product.ts
@@ -8,7 +8,7 @@ export class QtiProduct extends QtiExpression<number> {
     const product = values.reduce((accumulator, currentValue) => {
       if (currentValue.baseType == 'float' || currentValue.baseType == 'integer') {
         try {
-          return accumulator * parseInt(currentValue.value.toString());
+          return accumulator * parseFloat(currentValue.value.toString());
         } catch (error) {
           console.warn(`can not convert to number`);
         }
